Split dom utils in smaller modules
